I redid the class Handler

+++ src/main/java/com/example/transport/service/TransportServiceImpl.java
@@ -16,18 +16,25 @@ public class TransportServiceImpl {
     public List<Transport> getAllTransport(){
         return transportRepo.findAll();
     }
+
     public void addNewTransport(Transport transport){
         transportRepo.save(transport);
     }
+
     public Transport getTransportById(int id){
         return transportRepo.findById(id).orElseThrow(()->new TransportException("Transport with id " + id + " was not found"));
     }
+
     public void updateTransport(int id, Transport transport) {
         Transport oldTransport = transportRepo.getReferenceById(id);
         oldTransport.setId(transport.getId());
         transportRepo.save(oldTransport);
     }
+
     public void deleteTransport(int id){
+        if(transportRepo.existsById(id)){
+            throw new TransportException("Transport with id " + id + " not found");
+        }
         transportRepo.deleteById(id);
     }
 }
